Integration test: Verify finish state & add auto-rerun in regenerate.sh (#1773)
* regenerate.sh: Allow testing on a specific agent and/or test

* Check agent finish state

* rengerate.sh: Rerun after fixing the prompts

* Fix SWEAgent test_write_simple_script

* Add more help message

* Add a known issue to README.md

* regenerate.sh: Fix help message typo

* Fix a typo in README

+++ tests/integration/regenerate.sh
@@ -1,6 +1,18 @@
 #!/bin/bash
 set -eo pipefail
 
+run_test() {
+  SANDBOX_TYPE=$SANDBOX_TYPE \
+    WORKSPACE_BASE=$WORKSPACE_BASE \
+    REMIND_ITERATIONS=$remind_iterations \
+    MAX_ITERATIONS=$MAX_ITERATIONS \
+    WORKSPACE_MOUNT_PATH=$WORKSPACE_MOUNT_PATH \
+    AGENT=$agent \
+    poetry run pytest -s ./tests/integration/test_agent.py::$test_name
+    # return exit code of pytest
+    return $?
+}
+
 if [ -z $WORKSPACE_MOUNT_PATH ]; then
   WORKSPACE_MOUNT_PATH=$(pwd)
 fi
@@ -30,14 +42,35 @@ test_names=(
 num_of_tests=${#test_names[@]}
 num_of_agents=${#agents[@]}
 
+if [ "$num_of_agents" -ne "${#remind_iterations_config[@]}" ]; then
+  echo "Every agent must have its own remind_iterations_config"
+  exit 1
+fi
+
+if [ "$num_of_tests" -ne "${#test_names[@]}" ]; then
+  echo "Every task must correspond to one test case"
+  exit 1
+fi
+
 rm -rf logs
 rm -rf $WORKSPACE_BASE
 for ((i = 0; i < num_of_tests; i++)); do
   task=${tasks[i]}
   test_name=${test_names[i]}
+
+  # skip other tests if only one test is specified
+  if [[ -n "$ONLY_TEST_NAME" && "$ONLY_TEST_NAME" != "$test_name" ]]; then
+    continue
+  fi
+
   for ((j = 0; j < num_of_agents; j++)); do
     agent=${agents[j]}
     remind_iterations=${remind_iterations_config[j]}
+
+    # skip other agents if only one agent is specified
+    if [[ -n "$ONLY_TEST_AGENT" && "$ONLY_TEST_AGENT" != "$agent" ]]; then
+      continue
+    fi
 
     echo -e "\n\n\n\n========Running $test_name for $agent========\n\n\n\n"
     rm -rf $WORKSPACE_BASE
@@ -53,13 +86,7 @@ for ((i = 0; i < num_of_tests; i++)); do
       set +e
     fi
 
-    SANDBOX_TYPE=$SANDBOX_TYPE \
-      WORKSPACE_BASE=$WORKSPACE_BASE \
-      REMIND_ITERATIONS=$remind_iterations \
-      MAX_ITERATIONS=$MAX_ITERATIONS \
-      WORKSPACE_MOUNT_PATH=$WORKSPACE_MOUNT_PATH \
-      AGENT=$agent \
-      poetry run pytest -s ./tests/integration/test_agent.py::$test_name
+    run_test
     TEST_STATUS=$?
     # Re-enable 'exit on error'
     set -e
@@ -93,6 +120,31 @@ for ((i = 0; i < num_of_tests; i++)); do
 
       mkdir -p tests/integration/mock/$agent/$test_name/
       mv logs/llm/**/* tests/integration/mock/$agent/$test_name/
+
+      echo -e "\n\n\n\n========$test_name test data regenerated for $agent, rerun test again to verify========\n\n\n\n"
+      # Temporarily disable 'exit on error'
+      set +e
+      run_test
+      TEST_STATUS=$?
+      # Re-enable 'exit on error'
+      set -e
+
+      if [[ $TEST_STATUS -ne 0 ]]; then
+        echo -e "\n\n\n\n========$test_name for $agent RERUN FAILED========\n\n\n\n"
+        echo -e "There are multiple possibilities:"
+        echo -e "  1. The agent is unable to finish the task within $MAX_ITERATIONS steps."
+        echo -e "  2. The agent thinks itself has finished the task, but fails the validation in the test code."
+        echo -e "  3. There is something non-deterministic in the prompt."
+        echo -e "  4. There is a bug in this script, or in OpenDevin code."
+        echo -e "NOTE: Some of the above problems could sometimes be fixed by a retry (with a more powerful LLM)."
+        echo -e "      You could also consider improving the agent, increasing MAX_ITERATIONS, or skipping this test for this agent."
+        exit 1
+      else
+        echo -e "\n\n\n\n========$test_name for $agent RERUN PASSED========\n\n\n\n"
+        sleep 1
+      fi
+
+
     else
       echo -e "\n\n\n\n========$test_name for $agent PASSED========\n\n\n\n"
       sleep 1
